Переписав всі Domain Entities від BaseEntity

+++ Domain/Admins/IAdminRepository.cs
@@ -4,11 +4,11 @@
     {
         void Create(Admin admin);
         Admin Update(Admin admin);
-        Admin GetByAdminId(int id);
+        Admin GetByAdminId(long id);
         Admin GetByRecoveryCode(int recoveryCode);
         Admin GetByEmail(string email);
         Admin GetByPasswordToken(string email);
-        Admin[] GetActiveAdmins(int adminId, int since, int count);
+        Admin[] GetActiveAdmins(long adminId, int since, int count);
         dynamic GetUser(int since, int count);
         dynamic GetFollowers(int since, int count);
     }
