+++ Domain/Admins/IAppealRepository.cs
@@ -4,8 +4,8 @@
     {
         void Create(Appeal appeal);
         void Update(Appeal appeal);
-        Appeal GetBy(int appealId);
-        Appeal GetBy(int appealId, string userToken);
+        Appeal GetBy(long appealId);
+        Appeal GetBy(long appealId, string userToken);
         Appeal[] GetAppealsBy(int since, int count);
         Appeal[] GetAppealsBy(string userToken, int since = 0, int count = 10);
     }
